pyproject.toml
[project]
name = "tech-bridge-back"
version = "0.1.0"
description = ""
authors = [
    {name = "Malik Laet",email = "isklikma@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "fast_tech"}]



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"
ruff = "^0.11.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev fast_tech/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_tech -vv'
post_test = 'coverage html'